static cast of size and string_view for decision response

+++ bindings/cs/rl.net.native/rl.net.continuous_action_response.cc
@@ -13,14 +13,14 @@ API void DeleteContinuousActionResponse(reinforcement_learning::continuous_actio
 API const char* GetContinuousActionEventId(reinforcement_learning::continuous_action_response* response, int& event_id_size)
 {
     const auto event_id = response->get_event_id();
-    event_id_size = event_id.size();
+    event_id_size = static_cast<int>(event_id.size());
     return event_id.data();
 }
 
 API const char* GetContinuousActionModelId(reinforcement_learning::continuous_action_response* response, int& model_id_size)
 {
     const auto model_id = response->get_model_id();
-    model_id_size = model_id.size();
+    model_id_size = static_cast<int>(model_id.size());
     return model_id.data();
 }
